Ryu-Snort integration that works with flows
Previously, I'd been trying to get this to work by making modifications
to simple_switch_snort.py. However, the code to add the flow to the
switch was not working as expected. Whenever I had this code uncommented,
the Snort server would not receive packets anymore. Commenting it out only
worked because it meant that all packets were still coming through the
controller and the Python code there was directly sending it to both its
destination and the Snort server. But obviously you need to install flows
so that the controller isn't a bottleneck for packets coming through.

I tried briefly to debug before noticing that simple_switch_13.py had
more nuanced handling surrounding the flow installation. This commit ports
the Snort integration to that file. I verified that installed flow properly
mirrors packets to the Snort server.